Put default webpack config in dedicated package
