+++ packages/tooling/openmrs/package.json
@@ -6,6 +6,7 @@
   "bin": {
     "openmrs": "./dist/cli.js"
   },
+  "types": "src/index.ts",
   "engines": {
     "node": ">= 12.0.0"
   },
@@ -31,23 +32,16 @@
   ],
   "homepage": "https://github.com/openmrs/openmrs-esm-core#readme",
   "dependencies": {
-    "@babel/core": "^7.11.4",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-typescript": "^7.10.4",
     "@openmrs/esm-app-shell": "3.2.0",
+    "@openmrs/webpack-config": "^3.2.0",
     "@types/react": "^16.9.46",
     "@types/systemjs": "^6.1.0",
     "autoprefixer": "^10.4.2",
     "axios": "^0.21.1",
-    "babel-loader": "^8.2.2",
     "browserslist-config-openmrs": "^1.0.1",
-    "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^10.0.0",
-    "css-loader": "^5.2.4",
     "cssnano": "^5.0.16",
     "ejs": "^2.6.2",
-    "fork-ts-checker-webpack-plugin": "^6.5.0",
     "glob": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
     "inquirer": "^7.3.3",
@@ -55,19 +49,13 @@
     "postcss": "^8.4.6",
     "postcss-loader": "^6.2.1",
     "rimraf": "^3.0.2",
-    "sass": "^1.44.0",
-    "sass-loader": "^12.3.0",
-    "style-loader": "^3.3.1",
-    "systemjs-webpack-interop": "^2.3.7",
     "tar": "^6.0.5",
     "ts-loader": "^9.2.6",
     "typescript": "~4.5.2",
     "webpack": "^5.64.0",
-    "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.6.0",
     "webpack-pwa-manifest": "^4.3.0",
-    "webpack-stats-plugin": "^1.0.3",
     "workbox-webpack-plugin": "^6.4.1",
     "yargs": "16.0.3"
   },
