+++ packages/tooling/webpack-config/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "@openmrs/webpack-config",
+  "version": "3.2.0",
+  "license": "MPL-2.0",
+  "main": "dist/index.js",
+  "types": "src/index.ts",
+  "engines": {
+    "node": ">= 12.0.0"
+  },
+  "scripts": {
+    "test": "jest --passWithNoTests",
+    "build": "tsc",
+    "lint": "eslint src --ext ts,tsx"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/openmrs/openmrs-esm-core.git"
+  },
+  "bugs": {
+    "url": "https://github.com/openmrs/openmrs-esm-core/issues"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "keywords": [
+    "openmrs",
+    "microfrontends",
+    "webpack",
+    "config"
+  ],
+  "homepage": "https://github.com/openmrs/openmrs-esm-core#readme",
+  "dependencies": {
+    "@babel/core": "^7.11.4",
+    "@babel/preset-env": "^7.11.0",
+    "@babel/preset-react": "^7.10.4",
+    "@babel/preset-typescript": "^7.10.4",
+    "babel-loader": "^8.2.2",
+    "clean-webpack-plugin": "^4.0.0",
+    "css-loader": "^5.2.4",
+    "fork-ts-checker-webpack-plugin": "^6.5.0",
+    "lodash": "^4.17.21",
+    "sass": "^1.44.0",
+    "sass-loader": "^12.3.0",
+    "style-loader": "^3.3.1",
+    "systemjs-webpack-interop": "^2.3.7",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-stats-plugin": "^1.0.3"
+  },
+  "peerDependencies": {
+    "webpack": "5.x"
+  },
+  "devDependencies": {
+    "webpack": "^5.64.0",
+    "typescript": "~4.5.2"
+  }
+}
